move websocket outside
