+++ app/middlewares/index.js
@@ -1,0 +1,4 @@
+export { default as logger } from './logger'
+export { default as navigation } from './navigation'
+export { default as persistence } from './persistence'
+export { default as websocket } from './websocket'
