+++ app/middlewares/websocket/index.js
@@ -1,3 +1,1 @@
-import middleware from './middleware'
-
-export default middleware
+export { default } from './middleware'
